Fixed typos. Almost works.
Rolled back to previuos program model, fixed typos.
Now monkey grabs the banana almost correctly. But as result of comparison with the most goal structure a new undefined variable appears (clearly seen in the trace list).

Tip: DO NOT code at evening time, do it in the morning or... night :).